+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Session log now shows the version of tasmota-webserial-esptool package along with a timestamp
 - Failed detection of flashid is now non-fatal
 
+### Fixed
+- Ensure ESP32 SPIFFS headers reserve the default 4 bytes of metadata to match the ESP-IDF layout related to ([issue #77](https://github.com/thelastoutpostworkshop/ESPConnect/issues/77)).
+
 ### Internal
 - Serial Monitor start now use native tasmota-webserial-esptool hardreset(false) to enter user firmware
 - Bumps tasmota-webserial-esptool from 7.2.2 to 7.2.3